% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce cols.R
\name{coalesce_char}
\alias{coalesce_char}
\title{Coalesce columns as characters}
\usage{
coalesce_char(...)
}
\arguments{
\item{...}{Vectors}
}
\value{
A character vector
}
\description{
Takes any number of vectors as input, coerces to character vectors and
uses dplyr::coalesce to coalesce them
}
