% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce cols.R
\name{coalesce_factor}
\alias{coalesce_factor}
\title{Coalesce factor columns}
\usage{
coalesce_factor(...)
}
\arguments{
\item{...}{Vectors}
}
\value{
A factor vector
}
\description{
Takes any number of factor vectors as input, coerces them into charactor vectors.
Coerces charactor vector into factor with combined levels of input vectors.
}
