% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce cols.R
\name{combine_levels}
\alias{combine_levels}
\title{Combine levels}
\usage{
combine_levels(...)
}
\arguments{
\item{...}{Vectors}
}
\value{
A character vector
}
\description{
Takes any number of vectors as input, finds all unique levels of all
input vectos
}
