% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{factor_relevel}
\alias{factor_relevel}
\title{Relevels categorical variable to factor}
\usage{
factor_relevel(x, levels_list)
}
\arguments{
\item{x}{A categorical (numerical or character) variable}

\item{levels_list}{List of matched value pairs eg list('1' = 'mand', '2' = 'kvinde', etc)}
}
\value{
Factor with specified levels
}
\description{
Create factor from categorical variable with values stored as eg. numbers
(1, 4, 5, 99, etc) to new (actual) levels (man, woman, unknown). Input the numerical
or character vector and list of matched old value = new value pairs.
}
