% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{gfunk.yearweek2date}
\alias{gfunk.yearweek2date}
\title{Convert ISO year-week to date}
\usage{
gfunk.yearweek2date(date)
}
\arguments{
\item{date}{Converts character string or vector in format "2020-W34" or "2020-24"}
}
\description{
Converts character string or vector in format "2020-W34" or "2020-24" to
thursday date in week
}
