% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic functions.R
\name{if_else_NA}
\alias{if_else_NA}
\title{Return 'true' or NA}
\usage{
if_else_NA(condition, true)
}
\arguments{
\item{condition}{A condition, evaluating to a logical value}

\item{true}{value to return if true}
}
\value{
true or NA of same class as true
}
\description{
Evaluates 'condition' and returns 'true' if TRUE and NA if FALSE
}
