% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_dictionary}
\alias{join_dictionary}
\title{Join dataframe with dictionary}
\usage{
join_dictionary(.data, dict, by = NULL, suffix = ".dict")
}
\arguments{
\item{.data}{A dataframe/tibble}

\item{dict}{A dataframe or tibble with 2 columns}

\item{by}{a one length named character vector specifying which columns to join by}

\item{suffix}{a character to append to new column name}
}
\value{
A joined dataframe
}
\description{
Joins dictionary with dataframe and renames dictionary variable. Dictionary is a tibble or
dataframe with two columns: one with values matching the specified column in the supplied
dataframe, and one with the "new" or translated values". After joining, the column with the
translated values will be renamed to the name of the old column with a suffix.
}
