% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic functions.R
\name{NA_class}
\alias{NA_class}
\title{Return NA of same Class as input}
\usage{
NA_class(x)
}
\arguments{
\item{x}{An object}
}
\value{
NA of same class as object
}
\description{
Return NA of same Class as input
}
