% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scale_fill_super}
\alias{scale_fill_super}
\title{Scale fill with SUPE_R}
\usage{
scale_fill_super(
  palette = "cool",
  discrete = TRUE,
  reverse = FALSE,
  n_colors = 16,
  ...
)
}
\arguments{
\item{discrete}{Logical}

\item{reverse}{Reverse palette}

\item{pallette}{Palette to create color ramp from}
}
\description{
Scale fill with SUPE_R
}
