% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{super_pal}
\alias{super_pal}
\title{Return SUPE_R themed palette}
\usage{
super_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{pallette}{Palette to create color ramp from}
}
\value{
list of color names
}
\description{
Return SUPE_R themed palette
}
